backend/requirements.txt
azure-core==1.39.0
azure-storage-blob==12.28.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==46.0.7
idna==3.11
isodate==0.7.2
numpy==2.4.4
pandas==3.0.2
pycparser==3.0
pyodbc==5.3.0
python-dateutil==2.9.0.post0
redis==7.4.0
requests==2.33.1
six==1.17.0
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
